requirements.txt
google-cloud-aiplatform
google-cloud-vision
pandas
python-dotenv
fastapi
uvicorn
python-multipart
